modification

+++ 09.program.c
@@ -6,21 +6,21 @@ int main ()
 {
     
     float num=20,num2=100;
-    int *num_ptr;
+    float *num_ptr;
     num_ptr=&num;
-    printf("The value of num is %d \n ",num);
-    printf("The address of num is %d \n",&num);
-    printf ("The value of the address of the integer data type is %d \n",num_ptr);
-    printf("De_referencing of pointer value is %d \n",*num_ptr);
+    printf("The value of num is %f \n ",num);
+    printf("The address of num is %f \n",&num);
+    printf ("The value of the address of the integer data type is %f \n",num_ptr);
+    printf("De_referencing of pointer value is %f \n",*num_ptr);
 
     *num_ptr=50;
-    printf("The value of num is %d \n ",num);
-    printf ("The value of the address of the integer data type is %d \n",num_ptr);
-    printf("De_referencing of pointer value is %d \n",*num_ptr);
+    printf("The value of num is %f \n ",num);
+    printf ("The value of the address of the integer data type is %f \n",num_ptr);
+    printf("De_referencing of pointer value is %f \n",*num_ptr);
 
     *num_ptr=&num2;
-    printf("The address of num2 is %d \n",*num_ptr);
-    printf("Dereferencing num 2 %d \n",*num_ptr);
+    printf("The address of num2 is %f \n",*num_ptr);
+    printf("Dereferencing num 2 %f \n",*num_ptr);
     printf("The value of num2 is %f \n ",num2);
